2_data-engineering-zoomcamp/4_prefect/flows/03_deployments/docker-requirements.txt
pandas
prefect-sqlalchemy
prefect-gcp[cloud_storage]
protobuf
pyarrow
pandas-gbq